Re-structuring.
